==> nlp/1b-SA-Lexicons.ipynb <==
## Language Analysis of Alexithymic Discourse

<hr>

Alexithymic Language Project / raul@psicobotica.com / V2 release (sept 2020)

<hr>

## Sentiment Analysis Lexicons (Spanish)

- Multilingual Sentiment Project (Spanish) [link](https://sites.google.com/site/datascienceslab/projects/multilingualsentiment).
- AFINN-165 (English) [link](https://github.com/fnielsen/afinn). 
- AFINN-165-ES (my translation of AFINN-165). 


## Multilingual Sentiment Project (Spanish)
Lists of positive and negative words

import urllib.request
import numpy as np
import pandas as pd

msp_pos_words_path = "https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/lexicon/Multilingualsentiment_positive_words_es.txt"
msp_neg_words_path = "https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/lexicon/Multilingualsentiment_negative_words_es.txt"

# pos_text_file = open(msp_pos_words_path, "r")
# neg_text_file = open(msp_neg_words_path, "r")
# Load from URL instead of local disk
pos_text_file =  urllib.request.urlopen(msp_pos_words_path)
neg_text_file =  urllib.request.urlopen(msp_neg_words_path)

charset = pos_text_file.info().get_content_charset()
charset

# Format is one word per line
pos_lines = pos_text_file.read().decode(charset)
neg_lines = neg_text_file.read().decode(charset)

pos_lines = pos_lines.splitlines()
neg_lines = neg_lines.splitlines()

print("%2d %2d" % (len(neg_lines), len(pos_lines)))

neg_lines[34:40]

pos_lines[34:40]

# Build sets
pos_words_set = set(pos_lines)
neg_words_set = set(neg_lines)

pos_words_set.intersection(neg_words_set)

all_words_set = pos_words_set.union(neg_words_set)
len(all_words_set)

test_phrase = "Que asco de vida".split()

n_pos = np.sum([word in pos_words_set for word in test_phrase])
n_neg = np.sum([word in neg_words_set for word in test_phrase])

print("N_POS:%2d; N_NEG:%2d" % (n_pos, n_neg))

# Considering stems instead of full words

import nltk
from nltk.stem.porter import PorterStemmer
from nltk.stem import SnowballStemmer

# Porter stemmer
p_stemmer = PorterStemmer() 

# Snowball stemmer
s_stemmer = SnowballStemmer('spanish')

pos_stems_p = [p_stemmer.stem(word) for word in pos_lines]
neg_stems_p = [p_stemmer.stem(word) for word in neg_lines]

pos_stems_s = [s_stemmer.stem(word) for word in pos_lines]
neg_stems_s = [s_stemmer.stem(word) for word in neg_lines]

pos_stems_p_set = set(pos_stems_p)
neg_stems_p_set = set(neg_stems_p)

pos_stems_s_set = set(pos_stems_s)
neg_stems_s_set = set(neg_stems_s)


test_phrase2 = "Que asco de vida".split()
test_phrase2_s = [p_stemmer.stem(word) for word in test_phrase2]

n_pos = np.sum([word in pos_stems_s_set for word in test_phrase2_s])
n_neg = np.sum([word in neg_stems_s_set for word in test_phrase2_s])

print("N_POS:%2d; N_NEG:%2d" % (n_pos, n_neg))

## Create dataframes
pos_words_df = pd.DataFrame(list(pos_words_set), columns=['Pos'])
neg_words_df = pd.DataFrame(list(neg_words_set), columns=['Neg'])
pos_stems_s_df = pd.DataFrame(list(pos_stems_s_set), columns=['Pos'])
neg_stems_s_df = pd.DataFrame(list(neg_stems_s_set), columns=['Neg'])
pos_stems_p_df = pd.DataFrame(list(pos_stems_p_set), columns=['Pos'])
neg_stems_p_df = pd.DataFrame(list(neg_stems_p_set), columns=['Neg'])

pos_stems_s_df.count()

### Save word sets as df 

pos_words_df_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Sentiment_Lexicons\\MSP_Pos_Words.csv"
neg_words_df_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Sentiment_Lexicons\\MSP_Neg_Words.csv"
pos_stems_p_df_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Sentiment_Lexicons\\MSP_Pos_StemsP.csv"
neg_stems_p_df_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Sentiment_Lexicons\\MSP_Neg_StemsP.csv"
pos_stems_s_df_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Sentiment_Lexicons\\MSP_Pos_StemsS.csv"
neg_stems_s_df_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Sentiment_Lexicons\\MSP_Neg_StemsS.csv"

pos_words_df.to_csv(pos_words_df_path, sep=';', encoding='utf-8', index=False)
neg_words_df.to_csv(neg_words_df_path, sep=';', encoding='utf-8', index=False)
pos_stems_p_df.to_csv(pos_stems_p_df_path, sep=';', encoding='utf-8', index=False)
neg_stems_p_df.to_csv(neg_stems_p_df_path, sep=';', encoding='utf-8', index=False)
pos_stems_s_df.to_csv(pos_stems_s_df_path, sep=';', encoding='utf-8', index=False)
neg_stems_s_df.to_csv(neg_stems_s_df_path, sep=';', encoding='utf-8', index=False)

## AFINN-165

### Getting AFINN sentiment analysis lexicon in English

AFINN_path = "https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/lexicon/AFINN-en-165.txt"

AFINN_df = pd.read_csv(AFINN_path, header=None, delimiter="\t")

AFINN_df.columns = ['Word', 'Score']

AFINN_df.head()

AFINN_df.describe()

### Translate FINN words into Spanish
Using Google Translate API

# ! pip install googletrans

# Translating with Google Translate API
from googletrans import Translator 
translator = Translator()

AFINN_es_df = AFINN_df.copy()
AFINN_es_df['Word_ES_Text'] = ""

# Apply in batches 
for i in range(3000,3382):
    trlted = translator.translate(AFINN_df['Word'].iloc[i], src='en', dest='es').text
    print(trlted)
    AFINN_es_df['Word_ES_Text'].iloc[i] = trlted


# AFINN_es_df.head()

# All at once (problems with Google API)
# AFINN_df['Word_ES'] = AFINN_df.apply(
#     lambda row: translator.translate(row.Word, src='en', dest='es').text, axis=1) 

# Remove duplicates (they appeared due to translation)
AFINN_es_df = AFINN_es_df.drop_duplicates(subset='Word_ES_Text', keep="first")

len(AFINN_df) - len(AFINN_es_df)

# AFINN_es_df = pd.read_csv("D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Sentiment_Lexicons\\AFINN-165-es.csv", header=0, delimiter=";")

# All to lower
AFINN_es_df['Word_ES_Text_lower'] = AFINN_es_df.Word_ES_Text.str.lower()

AFINN_es_df.drop('Word_ES', axis=1, inplace=True)

AFINN_es_df.drop('Word_ES_Text', axis=1, inplace=True)

AFINN_es_df = AFINN_es_df.rename(columns={'Word_ES_Text_lower': 'Word_ES'})

AFINN_es_df.head()

p_stemmer.stem('abandonar')

# Add Porter stems
AFINN_es_df['Stem_ES_P'] = AFINN_es_df.Word_ES.apply(lambda x: p_stemmer.stem(x) )

# Add Snowball stems
AFINN_es_df['Stem_ES_S'] = AFINN_es_df.Word_ES.apply(lambda x: s_stemmer.stem(x) )

AFINN_es_df.head(10)

### Saving Spanish Traslated AFINN df

AFINN_es_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Sentiment_Lexicons\\AFINN-165-es.csv"
AFINN_es_df.to_csv(AFINN_es_path, sep=';', encoding='utf-8', index=False)
