==> nlp/2_Features.ipynb <==
## Language Analysis of Alexithymic Discourse

<hr>

Alexithymic Language Project / raul@psicobotica.com / V2 release (sept 2020)

<hr>

### NLP Feature Engineering

- Pre-processed Dataset load (already tokenized, stemmed, POS, NER, Lex, Dep parsed). 
- Cuantitative variables: counts (chars, words, sentences, punctuation, stopwords, title words, etc.)
- Cuantitative variables: average lengths.
- Cuantitative variables: POS frequencies.
- Cuantitative variables: diversity scores (HTR, TTR).

### Pre-Processed Dataset Load

import pandas as pd 

processed_dataset_path = "https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/data/Prolexitim_v2_processed_dep.csv"

alex_df = pd.read_csv(processed_dataset_path, header=0, delimiter=";")

alex_df.Code.count()

alex_df.columns

alex_df.head()

# Cuantitative NLP features

## Counts

### Char count (Chars)

alex_df['Chars'] = alex_df.Text.apply(lambda x: sum(len(word) for word in str(x).split(" ")))

alex_df.Chars.describe()

### Average word length (avgWL)

alex_df['avgWL'] = alex_df.Chars / alex_df.Words

alex_df.avgWL.describe()

### Average sentence length (avgSL)

alex_df['avgSL'] = alex_df.Words / alex_df.Sentences

alex_df.avgSL.describe()

### Puctuation Count (Pun_Count)

import string

count = lambda l1, l2: len(list(filter(lambda c: c in l2, l1)))

alex_df['Pun_Count'] = alex_df.Text.apply(lambda text: count(text,set(string.punctuation))) 

alex_df.Pun_Count.describe()

### Stop Words Count (Stop_Count)

# nltk stop words for Spanish doesn't seem sufficient
# ! pip install stop-words

import ast

import nltk
from nltk.corpus import stopwords
from stop_words import get_stop_words

# Stop words in Spanish 
es_stop = list(stopwords.words('spanish'))
es_stop2 = list(get_stop_words('spanish'))
es_stop.extend(es_stop2)

# len([w for w in ['hola','que','haces','perro',';','dia','sdfs','es'] if w in es_stop])

alex_df['Stop_Count'] = alex_df.Tokens.apply(lambda tokens: len([w for w in ast.literal_eval(tokens) if w in es_stop])) 

alex_df.Stop_Count.describe()

### Title Words Count (Title_Count)
A title word is defined as any word starting with a capital letter.

# I'd need to add the list of tokens without lower()
import nltk
from nltk.tokenize import RegexpTokenizer 
tokenizer = RegexpTokenizer(r'\w+')
alex_df['RawTokens'] = alex_df.apply(lambda row: tokenizer.tokenize(row.Text), axis=1)

alex_df['Title_Count'] = alex_df.RawTokens.apply(lambda tokens: len([w for w in tokens if w[0].isupper()])) 

alex_df.Title_Count.describe()

### Upper Case char count (Upper_Count)
Number of chars capitalized wihtin the text

alex_df['Upper_Count'] = alex_df.Text.apply(lambda text: sum(1 for c in text if c.isupper())) 

alex_df.Upper_Count.describe()

## PoS Counts

### Number of POS tags
VERB_Count<br>
NOUN_Count<br>
SYM_Count<br>
ADV_Count	<br>
PUNCT_Count<br>
INTJ_Count<br>
CCONJ_Count<br>
ADJ_Count<br>
AUX_Count<br>
DET_Count<br>
SCONJ_Count<br>
PRON_Count<br>
NUM_Count<br>
PROPN_Count<br>
ADP_Count<br>

# Get list of possible POS tags 
POS_tags = set()

for str_posList in alex_df.POS:
    posList = ast.literal_eval(str_posList)
    for posPair in posList:
        POS_tags.add(posPair[1])
# POS_tags

# Create df with POS tag counts
cNames = [] # Count column name
tNames = [] # Original tag name

for tag in POS_tags:
    cNames.append(tag+'_Count')
    tNames.append(tag)
    
POS_Counts_df = df = pd.DataFrame(columns = cNames)
POS_Counts_df

# for each exemplar, for each POS tag list, for each tag: count

for str_posList in alex_df.POS:
    posList = ast.literal_eval(str_posList)
    counts = []
    for tag in tNames:
        tagCount = 0        
        for posPair in posList:
            if posPair[1] == tag:
                tagCount += 1
        counts.append(tagCount)
    row = pd.Series(counts, index=POS_Counts_df.columns)
    POS_Counts_df = POS_Counts_df.append(row, ignore_index=True)


POS_Counts_df.head()

# Concat to the main df (append new columns)
POS_Counts_df = POS_Counts_df.reset_index(drop=True)
alex_df = alex_df.reset_index(drop=True)

feats_df = pd.concat([alex_df,POS_Counts_df],axis=1)

feats_df.columns

###  Relative frequency of POS 
Proportion of each POS relative to the total number of tokens

for tag in tNames:
    feats_df[tag+'_Ratio'] = feats_df.apply(lambda row: row[tag+'_Count'] / row['Words'], axis=1)

feats_df.head()

## Lexical Diversity Scores
https://en.wikipedia.org/wiki/Lexical_diversity

### Type Token Ratio (TTR)
Calculates the overall complexity of the text based on the total number of unique word types used in the text.<br>
The total number of unique words divided by the total number of words in order to give the text a score from 0 to 1.<br>
This is a very text length sensible score. Not really significant for comparison if texts are not of the same length. 


def numberOfTypes( row ):
    num = 0
    for typeToken in cNames: 
        if row[typeToken] > 0: 
            num += 1
    return num

feats_df['TTR'] = feats_df.apply( lambda row: numberOfTypes(row) / len(ast.literal_eval(row.Tokens)), axis=1)

feats_df.TTR.describe()

### Hapax legomena/Token ratio (HTR)
The number of words that occur only once divided by the number of total words.

feats_df['HTR'] = feats_df.apply(
    lambda row: len(set(ast.literal_eval(row.Tokens))) / len(ast.literal_eval(row.Tokens)), axis=1)

feats_df.HTR.describe()

feats_df.columns

feats_df.head()

## Save features df

features_dataset_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Prolexitim-Dataset\\Prolexitim_v2_features.csv"
feats_df.to_csv(features_dataset_path, sep=';', encoding='utf-8', index=False)
