==> nlp/3c-N-Grams.ipynb <==
## Language Analysis of Alexithymic Discourse

<hr>

Alexithymic Language Project / raul@psicobotica.com / V2 release (sept 2020)

<hr>

### N-GRAMs Models

An N-Gram is "A contiguous sequence of N items from a given sample of text or speech".

- Char N-Grams.
- Word N-Grams.
- Language generation with N-Gram models. 

<hr>

[Explanation of N-Gram models](https://en.wikipedia.org/wiki/N-gram)

## Load features dataset
- Data is already pre-processed (1-Preprocessing). 
- Basic NLP features are already calculated (2-Features). 
- Some additional BoW features have been added (3-BoW).
- Some additional TF/IDF features have been added (3-TFIDF).

import pandas as pd 
import numpy as np
import ast
import heapq
import re
import string
import random
import nltk

feats_dataset_path = "https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/data/Prolexitim_v2_features_3.csv"
alex_df = pd.read_csv(feats_dataset_path, header=0, delimiter=";")

alex_df.columns

alex_df.head()

## Preparing the corpora
Let's get two corpora, one with "alexithymic language" and the other with "non-alexithymic language". 
- AlexDoc will contain merged text from TAS-20 positive users. 
- NoAlexDoc will contain merged text from TAS-20 negative users. 

# Get all the alexithymic lang together
AlexDoc = ' '.join(alex_df[alex_df.Alex_A == 1].Text)

# Get all non-alexithymic lang together
NoAlexDoc = ' '.join(alex_df[alex_df.Alex_A == 0].Text)

print("As expected, we have a quite unbalanced dataset: Alex: %d; NoAlex: %d." % (len(AlexDoc),len(NoAlexDoc)))

# AlexDoc

# All to lower case and remove punctuation
AlexDoc = AlexDoc.translate(str.maketrans('', '', string.punctuation)).lower()
NoAlexDoc = NoAlexDoc.translate(str.maketrans('', '', string.punctuation)).lower()

## Creating N-Gram models

### Char N-Gram Model

# Establish the length of the n-grams
char_ngram_length = 6

# Char n-gram calculation function. 
def compute_char_ngrams(doc, length):
    """
    Parameters
    ----------
    doc : str
        Document to extract n-grams from. 
    length : int
        Length of the n-grams (2 is bigram, 3 is a trigram, etc.)
    """
    
    char_ngrams = {}
    
    for i in range(len(doc)-length):  # For each char i in doc
        seq = doc[i:i+length]         # Get current sequence
        # print(seq)
        if seq not in char_ngrams.keys():       # Create entry for new sequences
            char_ngrams[seq] = []                  
        char_ngrams[seq].append(doc[i+length])  # Add new char
    
    return char_ngrams

# Compute char n-grams for the two corpora
alex_char_ngrams = compute_char_ngrams(AlexDoc, char_ngram_length)
noalex_char_ngrams = compute_char_ngrams(NoAlexDoc, char_ngram_length)

alex_char_ngrams.get("niño q")

noalex_char_ngrams.get("lloran")

### Persist the char n-gram models

import json

charNGramAlex_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Prolexitim-Dataset\\Prolexitim_v2_CharNGram_Alex.json"
charNGramNoAlex_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Prolexitim-Dataset\\Prolexitim_v2_CharNGram_NoAlex.json"

with open(charNGramAlex_path, 'w', encoding='utf8') as fp:
    json.dump(alex_char_ngrams, fp, ensure_ascii=False)
    
with open(charNGramNoAlex_path, 'w', encoding='utf8') as fp:
    json.dump(noalex_char_ngrams, fp, ensure_ascii=False)

### Predicting text with the Char N-Gram Model

# N-Gram based char generation function. 
def predict_chars(seed, ng_model, ng_length, gen_length):
    """
    Parameters
    ----------
    seed : str
        Initial sequence of chars to generate from. 
    ng_model : dict
        Dictionary with the char N-Gram model. 
    ng_length : int
        Length of the n-grams (2 is bigram, 3 is a trigram, etc.)
    gen_length : int
        Length of the string to be generated.        
    """
    if (len(seed) != ng_length):
        raise Exception("Seed length shoud be equal to N-Grams length.")
    
    generated = seed  # We start with the seed. 

    for i in range(gen_length):
        if seed not in ng_model.keys():
            break
        possible_chars = ng_model[seed]
        next_char = possible_chars[random.randrange(len(possible_chars))]
        generated += next_char
        seed = generated[len(generated)-ng_length:len(generated)] # New seed is now the last sequence.   
    
    return generated

# Speak in "alexithymic language" prediction 
predict_chars("carlit", alex_char_ngrams, char_ngram_length, 200)

# Speak in "non-alexithymic language" prediction 
predict_chars("carlit", noalex_char_ngrams, char_ngram_length, 200)

# Speak in "non-alexithymic language" prediction 
predict_chars("niño q", noalex_char_ngrams, char_ngram_length, 50)

### Word N-Gram Model

# Establish the length of the n-grams
# Let's go for word trigrams
word_ngram_length = 3

# Word n-gram calculation function. 
def compute_word_ngrams(doc, length):
    """
    Parameters
    ----------
    doc : str
        Document to extract n-grams from. 
    length : int
        Length of the n-grams (2 is bigram, 3 is a trigram, etc.)
    """
    ngrams = {}
    
    tokens = nltk.word_tokenize(doc)
    for i in range(len(tokens)-length):
        seq = ' '.join(tokens[i:i+length])
        # print(seq)
        if seq not in ngrams.keys():
            ngrams[seq] = []
        ngrams[seq].append(tokens[i+length])
        
    return ngrams

# Compute word n-grams for the two corpora
alex_word_ngrams = compute_word_ngrams(AlexDoc, word_ngram_length)
noalex_word_ngrams = compute_word_ngrams(NoAlexDoc, word_ngram_length)

alex_word_ngrams.get('largo día de')

noalex_word_ngrams.get('largo día de')

### Persist the word n-gram models

wordNGramAlex_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Prolexitim-Dataset\\Prolexitim_v2_WordNGram_Alex.json"
wordNGramNoAlex_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Prolexitim-Dataset\\Prolexitim_v2_WordNGram_NoAlex.json"

with open(wordNGramAlex_path, 'w', encoding='utf8') as fp:
    json.dump(alex_word_ngrams, fp, ensure_ascii=False)
    
with open(wordNGramNoAlex_path, 'w', encoding='utf8') as fp:
    json.dump(noalex_word_ngrams, fp, ensure_ascii=False)

### Predicting text with the Word N-Gram Model

# N-Gram based word generation function. 
def predict_words(seed, ng_model, ng_length, gen_length):
    """
    Parameters
    ----------
    seed : str
        Initial sequence of words to generate from. 
    ng_model : dict
        Dictionary with the word N-Gram model. 
    ng_length : int
        Length of the n-grams (2 is bigram, 3 is a trigram, etc.)
    gen_length : int
        Length of the string to be generated (in words).        
    """
    seed_words = nltk.word_tokenize(seed)
    
    if (len(seed_words) != ng_length):
        raise Exception("Seed length shoud be equal to N-Grams length.")
    
    generated = seed
    for i in range(gen_length):
        if seed not in ng_model.keys():
            break
        possible_words = ng_model[seed]
        next_word = possible_words[random.randrange(len(possible_words))]
        generated += ' ' + next_word
        seq_words = nltk.word_tokenize(generated)
        seed = ' '.join(seq_words[len(seq_words)-ng_length:len(seq_words)])
    
    return generated

# Generate alexithymic text
predict_words('largo día de', alex_word_ngrams, word_ngram_length, 50)

# Generate non-alexithymic text
predict_words('largo día de', noalex_word_ngrams, word_ngram_length, 50)

predict_words('un niño que', noalex_word_ngrams, word_ngram_length, 100)

predict_words('un niño que', noalex_word_ngrams, word_ngram_length, 100)

predict_words('un niño que', alex_word_ngrams, word_ngram_length, 100)
