==> nlp/5_Personality.ipynb <==
## Language Analysis of Alexithymic Discourse

<hr>

Alexithymic Language Project / raul@psicobotica.com / V2 release (sept 2020)

<hr>

### Personality traits detection

There are many mechanisms by which personality traits of the author might be inferred from their text. Here, instead of training our own model (which requires extensive annotated data), we use IBM Personality Insights API. 

- Establish connection with IBM Personality Insights API. 
- Load existing feature vectors and obtain texts by participant. 
- Participant personality profiling using the API. 
- Parsing variables for personality traits, needs, values and preferences. 
- Saving the new personality insisghts as additional feature dimensions. 

<hr>

- [IBM Personality Insights API](https://www.ibm.com/watson/services/personality-insights/)
- [Big Five Personality Facets](https://github.com/watson-developer-cloud/doc-tutorial-downloads/blob/master/personality-insights/Personality-Insights-Facet-Characteristics.pdf)
- [Big Five Personality Dimensions](https://github.com/watson-developer-cloud/doc-tutorial-downloads/blob/master/personality-insights/Personality-Insights-Dimension-Characteristics.pdf)
- [Consumer Consumption Preference Prediction](https://github.com/watson-developer-cloud/doc-tutorial-downloads/blob/master/personality-insights/Improving-Consumption-Preferences-Accuracy.pdf)


### Using the Watson Developer Cloud Python SDK
- [Python SDK](https://github.com/watson-developer-cloud/python-sdk)
- [API Docs for Python](https://cloud.ibm.com/apidocs/personality-insights?code=python)

Using my own IBM Cloud account. 

Installing IBM Cloud python SDK


# ! pip install --upgrade ibm-watson

### IBM Cloud Authentication and endpoint

# Following the IBM cloud official example at:
# https://github.com/watson-developer-cloud/python-sdk/blob/master/examples/personality_insights_v3.py 
import pandas as pd
import numpy as np
import json
import os
from os.path import join
from ibm_watson import PersonalityInsightsV3
import csv
from ibm_cloud_sdk_core.authenticators import IAMAuthenticator

# Create the API Key
# API key stored in local file apikey.json
with open('apikey.json') as f:
    apikeydata = json.load(f)

# apikeydata.get('PI_key')

# Create a Personality Insight Price Plan (Lite: 1000 calls per month free)
# Select Location. 
# Personality Insights-ro activated
# Service credentials: get it from the list of resources. 

# Using London location (the only one in Europe): 
# https://api.eu-gb.personality-insights.watson.cloud.ibm.com 
# My instance: https://api.eu-gb.personality-insights.watson.cloud.ibm.com/instances/8d60aa18-abce-43f2-8ee0-bfa00b8a1140 

# authenticator = IAMAuthenticator(apikeydata.get('apikey'))
authenticator = IAMAuthenticator(apikeydata.get('PI_key'))
personality_insights = PersonalityInsightsV3(
    version='2017-10-13',  # as per last version documented in API Docs
    authenticator=authenticator
)

personality_insights.set_service_url('https://api.eu-gb.personality-insights.watson.cloud.ibm.com/instances/8d60aa18-abce-43f2-8ee0-bfa00b8a1140')

# Let's first try the API with a simple text 
test_Text = "Aunque el estudio de la mente y su relación con el cerebro es un fenómeno con una larga trayectoria histórica, el enfoque moderno interdisciplinar denominado neurociencia cognitiva aparece en 1976 de la mano de Michael Gazzaniga y George Miller. Durante las cuatro últimas décadas la ciencia del cerebro ha experimentado muchos avances significativos y el área multidisciplinar de la neurociencia cognitiva también ha ido experimentando cambios. Inicialmente, la neurociencia cognitiva se entendía como la colaboración entre el estudio de la mente (psicología) y el estudio del sistema nervioso (neurociencias). En las primeras etapas, el objetivo primordial era entender cómo los procesos neurobiológicos daban lugar a los procesos cognitivos. Es decir, comprender la cognición a través de sus correlatos neurobiológicos. Sin embargo, la mente no se puede explicar completamente desde la dimensión cognitiva (procesamiento de la información), sino que, como bien sabemos hoy en día, la emoción juega también un papel primordial. Por este motivo, algunos autores empezaron a habla de neurociencia afectiva, para resaltar el hecho de que el estudio del sistema nervioso también ha de ir dirigido a comprender la dimensión emocional de la mente."
len(test_Text.split())

# Getting the results as CSV:
from ibm_watson import ApiException
try:
    # Invoke a Personality Insights method
    response = personality_insights.profile(
        test_Text,
        content_type='text/plain',
        content_language='es',
        accept_language='en',
        consumption_preferences=True,
        raw_scores=True,
        accept='text/csv',
        csv_headers=True).get_result()
except ApiException as ex:
    print("Method failed with status code " + str(ex.code) + ": " + ex.message)

profile = response.content
cr = csv.reader(profile.decode('utf-8').splitlines())
my_list = list(cr)
for row in my_list:
    print(row[0:10])

import io
df = pd.read_csv(io.StringIO(profile.decode('utf-8')), lineterminator='\n')

df

# Getting the results as JSON:
from ibm_watson import ApiException
try:
    # Invoke a Personality Insights method
    response = personality_insights.profile(
        test_Text,
        content_type='text/plain',
        content_language='es',
        accept_language='en',
        consumption_preferences=True,
        raw_scores=True,
        accept='application/json').get_result()
except ApiException as ex:
    print("Method failed with status code " + str(ex.code) + ": " + ex.message)

print(json.dumps(response, indent=2))

### Accessing results from Personality Insights profiling

for i in response['personality']:
    print(i['name'])

# Personality traits percentiles:
for i in range(len(response['personality'])):
    print(response['personality'][i]['trait_id'] + 
          " --> %f" % (response['personality'][i]['percentile']))
    for j in range(len(response['personality'][i]['children'])):
        print(" > " + response['personality'][i]['children'][j]['trait_id'] +
             " --> %f" % response['personality'][i]['children'][j]['percentile'])

# Personality traits raw scores:
for i in range(len(response['personality'])):
    print(response['personality'][i]['trait_id'] + 
          " --> %f" % (response['personality'][i]['raw_score']))
    for j in range(len(response['personality'][i]['children'])):
        print(" > " + response['personality'][i]['children'][j]['trait_id'] +
             " --> %f" % response['personality'][i]['children'][j]['raw_score'])

# Needs percentiles:
for i in range(len(response['needs'])):
    print(response['needs'][i]['trait_id'] + 
          " --> %f" % (response['needs'][i]['percentile']))

# Needs raw-scores:
for i in range(len(response['needs'])):
    print(response['needs'][i]['trait_id'] + 
          " --> %f" % (response['needs'][i]['raw_score']))

# Values percentiles:
for i in range(len(response['values'])):
    print(response['values'][i]['trait_id'] + 
          " --> %f" % (response['values'][i]['percentile']))

# Values raw score:
for i in range(len(response['values'])):
    print(response['values'][i]['trait_id'] + 
          " --> %f" % (response['values'][i]['raw_score']))

# Consumption Preferences
for i in range(len(response['consumption_preferences'])):
    print(response['consumption_preferences'][i]['consumption_preference_category_id'])
    for j in range(len(response['consumption_preferences'][i]['consumption_preferences'])):
        print(" > " + response['consumption_preferences'][i]['consumption_preferences'][j]['consumption_preference_id'] +
             " --> %.2f" % (response['consumption_preferences'][i]['consumption_preferences'][j]['score']))

## Gathering text from participants into individual documents for personality profiling

# Get features dataset
feats_dataset_path = "https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/data/Prolexitim_v2_features_4.csv"
alex_df = pd.read_csv(feats_dataset_path, header=0, delimiter=";")

alex_df.sample(2)

print("We have " + str(alex_df.Code.nunique()) + " unique participants.")

# Code is the anonymous unique participant indentifier
docs_df = alex_df[['Code', 'Text']].groupby('Code')['Text'].apply(" ".join).reset_index()

docs_df.info()

docs_df.Text[0]

docs_df.sample(4)

## Annotate personality traits for each participant

# This function calls the Personality Insights API and gets the profile
# for a given plain text in Spanish.
def get_Profile(text):
    """
    Parameters
    ----------
    text : str
        Document to be analyzed. 
    
    Returns
    -------
    json_profile: json object
        Personality profile
        
    """
    try:
        json_profile = personality_insights.profile(
            text,
            content_type='text/plain',
            content_language='es',
            accept_language='en',
            consumption_preferences=True,
            raw_scores=True,
            accept='application/json').get_result()
    except ApiException as ex:
        print("Method failed with status code " + str(ex.code) + ": " + ex.message)
        
    return json_profile

# Personality Insights variables
PI_columns = ['big5_openness','facet_adventurousness','facet_artistic_interests','facet_emotionality',
              'facet_imagination','facet_intellect','facet_liberalism', 'big5_conscientiousness',
              'facet_achievement_striving','facet_cautiousness','facet_dutifulness',
              'facet_orderliness','facet_self_discipline','facet_self_efficacy','big5_extraversion',
              'facet_activity_level','facet_assertiveness','facet_cheerfulness','facet_excitement_seeking',
              'facet_friendliness','facet_gregariousness','big5_agreeableness','facet_altruism',
              'facet_cooperation','facet_modesty','facet_morality','facet_sympathy','facet_trust',
              'big5_neuroticism','facet_anger','facet_anxiety','facet_depression',
              'facet_immoderation','facet_self_consciousness','facet_vulnerability',
              'need_challenge','need_closeness','need_curiosity','need_excitement',
              'need_harmony','need_ideal','need_liberty','need_love','need_practicality',
              'need_self_expression','need_stability','need_structure',
              'value_conservation','value_openness_to_change','value_hedonism',
              'value_self_enhancement','value_self_transcendence',
              'consumption_preferences_automobile_ownership_cost',
              'consumption_preferences_automobile_safety',
              'consumption_preferences_clothes_quality',
              'consumption_preferences_clothes_style',
              'consumption_preferences_clothes_comfort',
              'consumption_preferences_influence_brand_name',
              'consumption_preferences_influence_utility',
              'consumption_preferences_influence_online_ads',
              'consumption_preferences_influence_social_media',
              'consumption_preferences_influence_family_members',
              'consumption_preferences_spur_of_moment',
              'consumption_preferences_credit_card_payment',
              'consumption_preferences_health_and_activity',
              'consumption_preferences_eat_out',
              'consumption_preferences_gym_membership',
              'consumption_preferences_outdoor',
              'consumption_preferences_concerned_environment',
              'consumption_preferences_start_business',
              'consumption_preferences_movie_romance',
              'consumption_preferences_movie_adventure',
              'consumption_preferences_movie_horror',
              'consumption_preferences_movie_musical',
              'consumption_preferences_movie_historical',
              'consumption_preferences_movie_science_fiction',
              'consumption_preferences_movie_war',
              'consumption_preferences_movie_drama',
              'consumption_preferences_movie_action',
              'consumption_preferences_movie_documentary',
              'consumption_preferences_music_rap',
              'consumption_preferences_music_country',
              'consumption_preferences_music_r_b',
              'consumption_preferences_music_hip_hop',
              'consumption_preferences_music_live_event',
              'consumption_preferences_music_playing',
              'consumption_preferences_music_latin',
              'consumption_preferences_music_rock',
              'consumption_preferences_music_classical',
              'consumption_preferences_read_frequency',
              'consumption_preferences_books_entertainment_magazines',
              'consumption_preferences_books_non_fiction',
              'consumption_preferences_books_financial_investing',
              'consumption_preferences_books_autobiographies',
              'consumption_preferences_volunteer']

len(PI_columns)

# Create new PI columns in the features df
for col in PI_columns:
    alex_df[col] = np.nan

# This function fills the feature dataframe with personality profile for participant identified by code
def add_PersonalityData(df, code, profile):
    """
    Parameters
    ----------
    df : dataframe
        Dataframe with features. 
    code : str
        Participant unique identifier. 
    profile: json
        JSON object with personality results for participant code.
    
    Returns
    -------
    Nothing. The dataframe is filled with extended features corresponding to personality traits. 
    """    
    
    # Add personality traits and their children traits raw scores
    for i in range(len(profile['personality'])):
        col_name_pers = profile['personality'][i]['trait_id']
        col_value_pers = profile['personality'][i]['raw_score']
        df.loc[ df['Code']==code, col_name_pers ] = col_value_pers
        
        for j in range(len(profile['personality'][i]['children'])):
            col_name_child = profile['personality'][i]['children'][j]['trait_id']
            col_value_child = profile['personality'][i]['children'][j]['raw_score']
            df.loc[ df['Code']==code, col_name_child ] = col_value_child
            
            
    # Add needs
    for i in range(len(profile['needs'])):
        col_name_need = profile['needs'][i]['trait_id']
        col_value_need = profile['needs'][i]['raw_score']
        df.loc[ df['Code']==code, col_name_need ] = col_value_need

        
    # Add values
    for i in range(len(profile['values'])):
        col_name_values = profile['values'][i]['trait_id']
        col_value_values = profile['values'][i]['raw_score']
        df.loc[ df['Code']==code, col_name_values ] = col_value_values 
        
        
    # Add consumption preferences
    for i in range(len(profile['consumption_preferences'])):
        for j in range(len(profile['consumption_preferences'][i]['consumption_preferences'])):
            col_name_prefs = profile['consumption_preferences'][i]['consumption_preferences'][j]['consumption_preference_id']
            col_value_prefs = profile['consumption_preferences'][i]['consumption_preferences'][j]['score']
            df.loc[ df['Code']==code, col_name_prefs ] = col_value_prefs 
    

### Call the API for each participant and add the profile to features df

for i in range(len(docs_df)):
    
    # The API requires a minimum of 100 words
    if ( len(docs_df['Text'].iloc[i].split()) > 100 ):
        
        # Get the results for user i:
        personality = get_Profile(docs_df['Text'].iloc[i])
    
        # Update the feature vectors with personality vbles
        add_PersonalityData(alex_df, docs_df['Code'].iloc[i], personality)

alex_df[PI_columns].describe()

alex_df[PI_columns].isna().sum()[0]

# Instances annotated - 60%
1 - (alex_df[PI_columns].isna().sum()[0] / alex_df.count()[0])

## Persist annotated personality data

# Save Updated features dataset
Feats_5_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Prolexitim-Dataset\\Prolexitim_v2_features_5.csv"
alex_df.to_csv(Feats_5_path, sep=';', encoding='utf-8', index=False)

# np.set_printoptions(threshold=np.inf)
for i in alex_df.columns:
    print(i)
