==> nlp/6_Sentiment.ipynb <==
## Language Analysis of Alexithymic Discourse

<hr>

Alexithymic Language Project / raul@psicobotica.com / V2 release (sept 2020)

<hr>

### Sentiment Analysis

We perform here different methods for sentiment analysis. Expressed sentiment variables might be use as part of the feature vectors for detecting alexithymia.

- Lexicon-based sentiment analysis. 
- Third-party API based sentiment analysis. 
- Discussion and caveats about training our own sentiment analysis model. 

<hr>

[More about Sentiment Analysis](https://en.wikipedia.org/wiki/Sentiment_analysis)


import pandas as pd 
import numpy as np
import ast

## Load features dataset
- Data is already pre-processed (1-Preprocessing). 
- Basic NLP features are already calculated (2-Features). 
- Some additional BoW features have been added (3-BoW).
- Some additional TF/IDF features have been added (3-TFIDF).
- N-Gram models have been generated (3-N-Grams). 
- PoS lists for each document identified (4-Lexicosemantics). 
- Writer's personality variables inferred (5-Personality).

feats_dataset_path = "https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/data/Prolexitim_v2_features_5.csv"
alex_df = pd.read_csv(feats_dataset_path, header=0, delimiter=";")

alex_df.sample(2)

## Lexicon-based sentiment Analysis

### Load sentiment lexicon models
- Previously generated (1b-SA-Lexicons). 
- We have lists of positives and negative words/stems obtained from Multilingual Sentiment Project. 
- We have my Spanish translation of AFINN-165 (AFINN-165-ES)

# Lexicon files are already available at Github: 
path_MSP_Pos_Words = 'https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/lexicon/MSP_Pos_Words.csv'
path_MSP_Neg_Words = 'https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/lexicon/MSP_Neg_Words.csv'
path_MSP_Pos_StemsS = 'https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/lexicon/MSP_Pos_StemsS.csv'
path_MSP_Neg_StemsS = 'https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/lexicon/MSP_Neg_StemsS.csv'
path_AFINN_165_ES = 'https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/lexicon/AFINN-165-es.csv'

# Get the lexicons in memory as dataframes
MSP_Pos_Words_df = pd.read_csv(path_MSP_Pos_Words, header=0, delimiter=";")
MSP_Neg_Words_df = pd.read_csv(path_MSP_Neg_Words, header=0, delimiter=";")
MSP_Pos_StemsS_df = pd.read_csv(path_MSP_Pos_StemsS, header=0, delimiter=";")
MSP_Neg_StemsS_df = pd.read_csv(path_MSP_Neg_StemsS, header=0, delimiter=";")
AFINN_165_ES_df = pd.read_csv(path_AFINN_165_ES, header=0, delimiter=";")

AFINN_165_ES_df.sample(4)

# Use sets in the case of MSP, where there is no specific score
MSP_Pos_Words_set = set(MSP_Pos_Words_df.Pos)
MSP_Neg_Words_set = set(MSP_Neg_Words_df.Neg)
MSP_Pos_StemsS_set = set(MSP_Pos_StemsS_df.Pos)
MSP_Neg_StemsS_set = set(MSP_Neg_StemsS_df.Neg)

import random
print(random.sample(MSP_Pos_Words_set, 4))
print(random.sample(MSP_Neg_Words_set, 4))
print(random.sample(MSP_Pos_StemsS_set, 4))
print(random.sample(MSP_Neg_StemsS_set, 4))

### Calculate sentimen per document
- Based on MSP word matches. 
- Based on MSP snowball stems matches. 
- Based on AFINN word matches. 
- Based on AFINN snowball stems matches. 

# Calculates the sentiment score based on MSP pos/neg words/stems and doc length
def get_MSP_Sent(doc, posWordsSet, negWordsSet):
    """
    Parameters
    ----------
    doc : list
        List of stopped tokens / snowballed stems extracted from a text. 
    posWordsSet : set
        Set of positive words
    
    Returns
    -------
    score: float
        (Positive matches - negative matches) / length of token list
        
    """
    score = 0
    
    for tok in doc: 
        if (tok in posWordsSet):
            score += 1
        if (tok in negWordsSet):
            score -= 1
            
    return (score / len(doc))    

alex_df.columns[0:40]

# Test function with words
test_tokens = ast.literal_eval(alex_df['Tokens_Stop'][159])
print(test_tokens)
get_MSP_Sent(test_tokens, MSP_Pos_Words_set, MSP_Neg_Words_set)

# Test function with stems
test_tokens = ast.literal_eval(alex_df['Tokens_Stem_S'][159])
print(test_tokens)
get_MSP_Sent(test_tokens, MSP_Pos_StemsS_set, MSP_Neg_StemsS_set)

# Now, add the MSP sentiment score to all examplars
alex_df['MSP_Words'] = alex_df.Tokens_Stop.apply(lambda x: get_MSP_Sent(ast.literal_eval(x), MSP_Pos_Words_set, MSP_Neg_Words_set))
alex_df['MSP_Stems'] = alex_df.Tokens_Stem_S.apply(lambda x: get_MSP_Sent(ast.literal_eval(x), MSP_Pos_StemsS_set, MSP_Neg_StemsS_set))

alex_df[['MSP_Words','MSP_Stems']].describe()

Let's do AFINN now: 
- Polarity: (positive scores - negative scores).
- Mean Intensity: mean(sum(positives),sum(abs(negatives))). 
- Max Intensity: max(sum(positives),sum(abs(negatives))).

# Calculates the sentiment score (polarity) based on AFINN-ES score
def get_AFINN_Polarity(doc, isStem, AFINN):
    """
    Parameters
    ----------
    doc : list
        List of stopped tokens / snowballed stems extracted from a text. 
    isStem: boolean
        True means stem, false indicates word. 
    AFINN : dataframe
        Dataframe with words, stems and their corresponding sentiment score.
    
    Returns
    -------
    score: int
        sum(score(Positive matches)) - sum(score(negative matches)))
        
    """
    score = 0
    
    if (isStem):
        for tok in doc: 
            points = AFINN[AFINN['Stem_ES_S'] == tok]['Score']
            if (len(points) == 1):
                # print("Word: " + tok + " -> " + str(points.item()))
                score += points.item()
    else: 
        for tok in doc: 
            points = AFINN[AFINN['Word_ES'] == tok]['Score']
            if (len(points) == 1):
                # print("Word: " + tok + " -> " + str(points.item()))
                score += points.item()
            
    return score

# Test function with words
test_tokens = ast.literal_eval(alex_df['Tokens_Stop'][69])
print(test_tokens)
get_AFINN_Polarity(test_tokens, False, AFINN_165_ES_df)

# Test function with stems
test_tokens = ast.literal_eval(alex_df['Tokens_Stem_S'][69])
print(test_tokens)
get_AFINN_Polarity(test_tokens, True, AFINN_165_ES_df)

# Calculates the sentiment mean intensity based on AFINN-ES score
def get_AFINN_Intensity(doc, isStem, AFINN):
    """
    Parameters
    ----------
    doc : list
        List of stopped tokens / snowballed stems extracted from a text. 
    isStem: boolean
        True means stem, false indicates word. 
    AFINN : dataframe
        Dataframe with words, stems and their corresponding sentiment score.
    
    Returns
    -------
    score: float
         mean(sum(positives),sum(abs(negatives))).
        
    """
    positives = 0
    negatives = 0
    
    if (isStem):
        for tok in doc: 
            points = AFINN[AFINN['Stem_ES_S'] == tok]['Score']
            if (len(points) == 1):
                score = points.item()
                # print("Word: " + tok + " -> " + str(score))
                if (score > 0):
                    positives += score
                else: 
                    negatives += abs(score)
    else: 
        for tok in doc: 
            points = AFINN[AFINN['Word_ES'] == tok]['Score']
            if (len(points) == 1):
                score = points.item()
                # print("Word: " + tok + " -> " + str(score))
                if (score > 0):
                    positives += score
                else: 
                    negatives += abs(score)
                            
    return ((positives+negatives)/2)

# Test function with words
test_tokens = ast.literal_eval(alex_df['Tokens_Stop'][69])
print(test_tokens)
get_AFINN_Intensity(test_tokens, False, AFINN_165_ES_df)

# Test function with stems
test_tokens = ast.literal_eval(alex_df['Tokens_Stem_S'][69])
print(test_tokens)
get_AFINN_Intensity(test_tokens, True, AFINN_165_ES_df)

# Calculates the sentiment maximum intensity based on AFINN-ES score
def get_AFINN_Max_Intensity(doc, isStem, AFINN):
    """
    Parameters
    ----------
    doc : list
        List of stopped tokens / snowballed stems extracted from a text. 
    isStem: boolean
        True means stem, false indicates word. 
    AFINN : dataframe
        Dataframe with words, stems and their corresponding sentiment score.
    
    Returns
    -------
    score: float
         max(sum(positives),sum(abs(negatives))).
        
    """
    positives = 0
    negatives = 0
    
    if (isStem):
        for tok in doc: 
            points = AFINN[AFINN['Stem_ES_S'] == tok]['Score']
            if (len(points) == 1):
                score = points.item()
                # print("Word: " + tok + " -> " + str(score))
                if (score > 0):
                    positives += score
                else: 
                    negatives += abs(score)
    else: 
        for tok in doc: 
            points = AFINN[AFINN['Word_ES'] == tok]['Score']
            if (len(points) == 1):
                score = points.item()
                # print("Word: " + tok + " -> " + str(score))
                if (score > 0):
                    positives += score
                else: 
                    negatives += abs(score)
                            
    if (positives >= negatives):
        return positives
    else:
        return negatives

# Test function with words
test_tokens = ast.literal_eval(alex_df['Tokens_Stop'][69])
print(test_tokens)
get_AFINN_Max_Intensity(test_tokens, False, AFINN_165_ES_df)

# Test function with stems
test_tokens = ast.literal_eval(alex_df['Tokens_Stem_S'][69])
print(test_tokens)
get_AFINN_Max_Intensity(test_tokens, True, AFINN_165_ES_df)

# Now, add the AFINN sentiment scores to all exemplars
alex_df['AFINN_Words_Pol'] = alex_df.Tokens_Stop.apply(lambda x: get_AFINN_Polarity(ast.literal_eval(x), False, AFINN_165_ES_df))
alex_df['AFINN_Stems_Pol'] = alex_df.Tokens_Stem_S.apply(lambda x: get_AFINN_Polarity(ast.literal_eval(x), True, AFINN_165_ES_df))
alex_df['AFINN_Words_Int'] = alex_df.Tokens_Stop.apply(lambda x: get_AFINN_Intensity(ast.literal_eval(x), False, AFINN_165_ES_df))
alex_df['AFINN_Stems_Int'] = alex_df.Tokens_Stem_S.apply(lambda x: get_AFINN_Intensity(ast.literal_eval(x), True, AFINN_165_ES_df))
alex_df['AFINN_Words_Max'] = alex_df.Tokens_Stop.apply(lambda x: get_AFINN_Max_Intensity(ast.literal_eval(x), False, AFINN_165_ES_df))
alex_df['AFINN_Stems_Max'] = alex_df.Tokens_Stem_S.apply(lambda x: get_AFINN_Max_Intensity(ast.literal_eval(x), True, AFINN_165_ES_df))

alex_df[['AFINN_Words_Pol',
         'AFINN_Stems_Pol','AFINN_Words_Int','AFINN_Stems_Int',
         'AFINN_Words_Max','AFINN_Stems_Max']].describe()

alex_df[['AFINN_Words_Pol',
         'AFINN_Stems_Pol','AFINN_Words_Int','AFINN_Stems_Int',
         'AFINN_Words_Max','AFINN_Stems_Max']].corr()

alex_df[['MSP_Stems','AFINN_Stems_Pol']].corr()

# We have 8 new features related to sentiment. 
alex_df.columns[len(alex_df.columns)-8:len(alex_df.columns)]

## Sentiment Analysis using IBM Cloud
- Watson Natural Language Understanding (NLU) API

[Watson NLU](https://www.ibm.com/es-es/cloud/watson-natural-language-understanding)

[API Guide - Python](https://cloud.ibm.com/apidocs/natural-language-understanding?code=python)

[Limited support for Spanish Language](https://cloud.ibm.com/docs/natural-language-understanding?topic=natural-language-understanding-language-support#spanish)

- Basically, Watson NLU provides a sentiment score for Spanish, but not the details about different emotions (for that, we'd need to translate first into English). 

import json
from ibm_watson import NaturalLanguageUnderstandingV1
from ibm_watson.natural_language_understanding_v1 import Features, EmotionOptions, SentimentOptions
from ibm_cloud_sdk_core.authenticators import IAMAuthenticator
from ibm_watson import ApiException

# Watson NLU endpoint for Europe
NLU_URL = 'https://gateway-lon.watsonplatform.net/natural-language-understanding/api'

# API key stored in local file apikey.json
with open('apikey.json') as f:
    apikeydata = json.load(f)

# apikeydata.get('NLU_key')

authenticator = IAMAuthenticator(apikeydata.get('NLU_key'))
natural_language_understanding = NaturalLanguageUnderstandingV1(
    version='2020-08-01',
    authenticator=authenticator
)

natural_language_understanding.set_service_url(NLU_URL)

# Testing the API
# Making the call for Emotion and Sentiment
response = natural_language_understanding.analyze(
    text='Probamos con una frase en español, aunque IBM dice que el idioma español no está soportado para las funciones de emociones, pero sí para las de análisis del sentimiento.',
    features=Features(emotion=EmotionOptions(),
                      sentiment=SentimentOptions())).get_result()

# Testing the API
# Making the call for Emotion and Sentiment
response = natural_language_understanding.analyze(
    text='dos palabras más',
    features=Features(sentiment=SentimentOptions())).get_result()

# Check the result:
print(json.dumps(response, indent=2))

response['sentiment']['document']['score']

Let's annotate our data with the sentiment score from IBM NLU

# This function calls the Watson NLU API and gets the sentiment score
# for a given plain text in Spanish.
def get_Watson_Sentiment(text_es):
    """
    Parameters
    ----------
    text_es : str
        Document to be analyzed in Spanish
    
    Returns
    -------
    score: float
        Document sentiment analysis results (sentiment score)
        
    """
    try:
        json_response = natural_language_understanding.analyze(
            text = text_es,
            features = Features(sentiment=SentimentOptions())).get_result()
        
    except ApiException as ex:
        return np.nan
        print("Method failed with status code " + str(ex.code) + ": " + ex.message)
        
    return json_response['sentiment']['document']['score']

# Apply for all rows: 
# alex_df['Watson_Sent'] = alex_df.Text.apply(lambda x: get_Watson_Sentiment(x))

# Create empty column
alex_df['Watson_Sent'] = np.nan

# Let do that iteratively
for i in range(len(alex_df)):
    
    # The API requires a minimum of XX words??
    if ( len(alex_df['Text'].iloc[i].split()) > 1 ):
        
        # Get the results for user i:
        score = get_Watson_Sentiment(alex_df['Text'].iloc[i])
        print("("+str(i)+") Watson said: " + str(score))
        
        # Update the feature vectors
        alex_df['Watson_Sent'].iloc[i] = score
        

alex_df.Watson_Sent.describe()

alex_df[['MSP_Stems','AFINN_Stems_Pol','Watson_Sent']].corr()

## Save fatures data in a new version of the CSV

# We have 9 new features related to sentiment. 
alex_df.columns[len(alex_df.columns)-9:len(alex_df.columns)]

# Save Updated features dataset
Feats_6_path = "D:\\Dropbox-Array2001\\Dropbox\\DataSets\\Prolexitim-Dataset\\Prolexitim_v2_features_6.csv"
alex_df.to_csv(Feats_6_path, sep=';', encoding='utf-8', index=False)
