==> nlp/10_Embeddings_USEM.ipynb <==
## Language Analysis of Alexithymic Discourse

<hr>

Alexithymic Language Project / raul@psicobotica.com / V2 release (sept 2020)

<hr>

## Latent Space Feature Engineering (embeddings)
### Using Google Universal Sentence Encoder - Multilingual Large 3

- TF Hub Pre-trained model: https://tfhub.dev/google/universal-sentence-encoder-multilingual-large/3 

### Load my dataset 
(After former feature engineering)

import pandas as pd 

feats_dataset_path = "https://raw.githubusercontent.com/raul-arrabales/alexithymic-lang/master/data/Prolexitim_v2_features.csv"

alex_df = pd.read_csv(feats_dataset_path, header=0, delimiter=";")

alex_df.columns

alex_df.head(2)

### Environment setup

%%capture
#@title Setup Environment
# Install the latest Tensorflow version.
!pip install tensorflow_text
!pip install bokeh
!pip install simpleneighbors[annoy]
!pip install tqdm

#@title Setup common imports and functions
import numpy as np
import os
import pandas as pd
import tensorflow.compat.v2 as tf
import tensorflow_hub as hub
from tensorflow_text import SentencepieceTokenizer
import sklearn.metrics.pairwise


# Using the Google Universal Sentence Enconder
# Multilingual version for Spanish support (large 3)
module_url = 'https://tfhub.dev/google/universal-sentence-encoder-multilingual/3' #@param ['https://tfhub.dev/google/universal-sentence-encoder-multilingual/3', 'https://tfhub.dev/google/universal-sentence-encoder-multilingual-large/3']

model = hub.load(module_url)

def embed_text(input):
  return model(input)

### Computing Text Embeddings

one_test = alex_df.RawTokens.iloc[4]
one_test

embedded_test = embed_text(one_test)
embedded_test.numpy()[0]

# Compute the embedding for each document in the dataset (as a one-dim numpy array)
alex_df['Embed_USEM'] = alex_df.RawTokens.apply(lambda tokens: embed_text(tokens).numpy()[0] ) 

alex_df.head(3)

# Export file
from google.colab import files

alex_df.to_csv('Prolexitim_v2_features-embed-usem.csv', sep=';', encoding='utf-8', index=False) 
files.download('Prolexitim_v2_features-embed-usem.csv')
